# docs/swagger.yaml
basePath: /api/v1
definitions:
  auth.ResponseModel:
    properties:
      error:
        type: string
      status:
        type: string
    type: object
  auth.SignInParams:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
  auth.SignInResponse:
    properties:
      token:
        type: string
    type: object
  auth.User:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a sample server film library server.
  title: Film Library App API
  version: "1.0"
paths:
...
  /auth/signIn:
    post:
      consumes:
      - application/json
      description: Login
      parameters:
      - description: login and password
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/auth.SignInParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.SignInResponse'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: SignIn
      tags:
      - Auth
  /auth/signUp:
    post:
      consumes:
      - application/json
      description: Create account
      parameters:
      - description: user data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/auth.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.ResponseModel'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: SignUp
      tags:
      - Auth
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"